requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
sentence-transformers
google-generativeai
python-dotenv
pydantic
jinja2
pymongo
motor
passlib[bcrypt]
PyJWT
python-multipart
pillow
pytesseract
pydantic[email]
